Relocate TextFieldWidget to widgets directoy
