+++ packages/airview-cms/src/features/cms/widgets/index.js
@@ -1,1 +1,2 @@
 export { DynamicWidget } from "./dynamic-widget";
+export { TextFieldWidget } from "./text-field-widget";
